ddrf_tutorial/requirements.txt
admin==0.0.1
asgiref==3.4.1
backcall==0.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
decorator==5.1.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.7
django-extensions==3.1.5
django-on-heroku==1.1.2
django-rest-framework==0.1.0
djangorestframework==3.12.4
httpie==2.5.0
idna==3.2
ipython==7.29.0
jedi==0.18.0
matplotlib-inline==0.1.3
owner==0.0.10
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.22
psycopg2-binary==2.9.3
ptyprocess==0.7.0
Pygments==2.10.0
PySocks==1.7.1
pytz==2021.1
requests==2.26.0
requests-toolbelt==0.9.1
sender==0.3
sqlparse==0.4.1
traitlets==5.1.1
urllib3==1.26.6
username==2.1.0
wcwidth==0.2.5
whitenoise==5.3.0
